Simplify stuff even further
